+++ client/app/detail/[id].tsx
