+++ client/app/page.tsx
@@ -1,0 +1,17 @@
+import Navbar from "../components/Navbar"
+import Footer from "../components/Footer"
+import ProductsPage from "./products/page"
+
+
+
+export default function HomePage() {
+  return (
+<div> 
+
+    <Navbar/>
+    <ProductsPage/>
+    <Footer/>
+
+</div>
+  )
+}
