+++ client/app/products/page.tsx
@@ -1,7 +1,6 @@
-// ruta de products
 
 import React from "react";
-import ProductList from "../src/components/ProductList/ProductList";
+import ProductList from "../../components/ProductList";
 
 const ProductsPage = () => {
     return (
