+++ client/components/Detail.tsx
