+++ client/components/SearchBar.tsx
