--- client/pages/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Home: React.FC = () => {
-  return (
-    <h1 className="text-3xl font-bold underline">
-      HOLA MUNDO!
-    </h1>
-  );
-};
-
-export default Home;
